* origin/mysql-trunk:
  Fixed build break after renaming of struct.
  Bug#20015132 ALTER TABLE FAILS TO CHECK IF TABLE IS CORRUPTED
  WL#6815 Adapt MySQL Cluster to 5.7
  WL#6815 Adapt MySQL Cluster to 5.7
  Post push fix for Bug#18948649
  Bug#19912474 - MYSQL_EMBEDDED FAILS WHEN LINKER SELECTS WRONG VERSION OF HANDLER CONSTRUCTOR
  Bug#19843808: DEADLOCK ON FLUSH TABLES WITH READ LOCK + SHOW SLAVE STATUS
  BUG 20029398 - GLOBAL VARIABLE GRANT_OPTION IS UNUSED
  Bug#18948649 : OUT OF FILE DESCRIPTORS - VARIOUS CRASHES WHEN                TMP TABLES FAIL
  Bug #19528825 "UNABLE TO PURGE A RECORD" WITH IBUF AND VARCHAR FIELDS CONTAINING SPACES
  Bug#19803418   ASSERTION: TABLE->QUIESCE == QUIESCE_NONE              IN ROW0QUIESCE.CC LINE 683
  Bug#19949915  IGNORABLE EVENTS DO NOT WORK AND ARE NOT TESTED
  BUG#20001250: RPL P_S: CHANGE ALL USAGE OF EXECUTE TO               APPLY IN THE TABLE NAMES
  Post-conversion fix.
  Bug#20032114: IMPROVE ERROR MESSAGE AFTER FATAL SIGNAL

+++ sql/ha_ndbcluster.h
@@ -352,7 +352,7 @@ static void set_tabname(const char *pathname, char *tabname);
   int ndb_err(NdbTransaction*);
 
   my_bool register_query_cache_table(THD *thd, char *table_key,
-                                     uint key_length,
+                                     size_t key_length,
                                      qc_engine_callback *engine_callback,
                                      ulonglong *engine_data);
 enum_alter_inplace_result
